+++ src/orquestra/sdk/_base/_api/_wf_run.py
@@ -25,6 +25,7 @@ from ...exceptions import (
 from ...schema import ir
 from ...schema.configs import ConfigName
 from ...schema.local_database import StoredWorkflowRun
+from ...schema.responses import WorkflowResult
 from ...schema.workflow_run import ProjectId, State
 from ...schema.workflow_run import TaskRun as TaskRunModel
 from ...schema.workflow_run import TaskRunId
@@ -351,16 +352,12 @@ class WorkflowRun:
             warnings.filterwarnings("ignore", category=VersionMismatch)
             return self._runtime.get_workflow_run_status(self.run_id)
 
-    def get_results(self, wait: bool = False) -> t.Sequence[t.Any]:
+    def get_results_serialized(self, wait: bool = False) -> t.Sequence[WorkflowResult]:
         """
-        Retrieves workflow results, as returned by the workflow function.
+        Retrieves workflow results in serialized form.
 
-        A workflow function is expected to return task outputs
-        (ArtifactFutures) or constants (10, "hello", etc.). This method returns values
-        of these. The order is dictated by the return statement in the workflow
-        function, for example `return a, b, c` means this function returns (a, b, c).
-        See also:
-        https://refactored-disco-d576cb73.pages.github.io/docs/runtime/guides/workflow-syntax.html
+        Result value is a sequence of WorkflowResult objects where each can be
+        deserialized separately
 
         Args:
             wait:  whether or not to wait for workflow run completion.
@@ -383,22 +380,61 @@ class WorkflowRun:
                 state,
             )
         try:
-            results = (
-                *(
-                    serde.deserialize(o)
-                    for o in self._runtime.get_workflow_run_outputs_non_blocking(
-                        self.run_id
-                    )
-                ),
-            )
+            return self._runtime.get_workflow_run_outputs_non_blocking(self.run_id)
         except WorkflowRunNotSucceeded:
             raise
+
+    def get_results(self, wait: bool = False) -> t.Sequence[t.Any]:
+        """
+        Retrieves workflow results, as returned by the workflow function.
+
+        A workflow function is expected to return task outputs
+        (ArtifactFutures) or constants (10, "hello", etc.). This method returns values
+        of these. The order is dictated by the return statement in the workflow
+        function, for example `return a, b, c` means this function returns (a, b, c).
+        See also:
+        https://docs.orquestra.io/docs/core/sdk/guides/workflow-syntax.html/workflow-syntax.html
+
+        Args:
+            wait:  whether or not to wait for workflow run completion.
+                   Uses the default options for waiting, use ``wait_until_finished()`` for
+                   more control.
+
+        Raises:
+            WorkflowRunNotFinished: when the workflow run has not finished and ``wait`` is
+                                   False
+            WorkflowRunNotSucceeded: when the workflow is no longer executing, but it did not
+                succeed.
+        """  # noqa 501
+        try:
+            serialized_results = self.get_results_serialized(wait=wait)
+        except WorkflowRunNotSucceeded:
+            raise
+
+        results = (*(serde.deserialize(o) for o in serialized_results),)
 
         # If we only get one result back, return it directly rather than as a sequence
         if len(results) == 1:
             return results[0]
 
         return results
+
+    def get_artifacts_serialized(
+        self,
+    ) -> t.Mapping[ir.TaskInvocationId, WorkflowResult]:
+        """
+        Unstable: this API will change.
+
+        Returns values calculated by this workflow's tasks in serialized form.
+        If a given task hasn't succeeded yet, the mapping won't
+        contain the corresponding entry.
+
+        Returns:
+            A dictionary with an entry for each task run in the workflow. The key is the
+                task's invocation ID. The value is whatever the task returned
+                in serialized form
+        """
+        return self._runtime.get_available_outputs(self.run_id)
 
     def get_artifacts(self) -> t.Mapping[ir.TaskInvocationId, t.Any]:
         """
@@ -415,7 +451,7 @@ class WorkflowRun:
         """
         # NOTE: this is a possible place for improvement. If future runtime APIs support
         # getting a subset of artifacts, we should use them here.
-        inv_outputs = self._runtime.get_available_outputs(self.run_id)
+        inv_outputs = self.get_artifacts_serialized()
 
         # The output shape differs across runtimes when the workflow functions returns a
         # single, packed future. See more in: